seperate findpeaks features from fpb
