+++ qt/widgets/common/inc/MantidQtWidgets/Common/FitPropertyBrowser.h
@@ -8,7 +8,6 @@
 
 #include "DllOption.h"
 #include "MantidAPI/AlgorithmObserver.h"
-#include "MantidAPI/IAlgorithm.h"
 #include "MantidAPI/Workspace_fwd.h"
 
 #include <QDockWidget>
@@ -20,9 +19,8 @@
 #include "MantidAPI/FunctionFactory.h"
 #include "MantidAPI/IFunction.h"
 #include "MantidAPI/IPeakFunction.h"
-#include "MantidAPI/ITableWorkspace.h"
 #include "MantidAPI/MatrixWorkspace_fwd.h"
-#include "MantidQtWidgets/Common/AlgorithmDialog.h"
+#include "MantidQtWidgets/Common/FitPropertyBrowserFindPeaksExt.h"
 #include "MantidQtWidgets/Common/IWorkspaceFitControl.h"
 #include "MantidQtWidgets/Common/SelectFunctionDialog.h"
 #include "MantidQtWidgets/Common/WorkspaceObserver.h"
@@ -52,72 +50,9 @@ class QPushButton;
 class QSignalMapper;
 class QTreeWidget;
 class QVBoxLayout;
-class QObject;
 
 namespace MantidQt {
 namespace MantidWidgets {
-
-class AlgorithmFinishObserver : public QObject, public Mantid::API::AlgorithmObserver {
-  Q_OBJECT
-signals:
-  /// Emitted when alg completes
-  void algCompletedSignal();
-
-protected:
-  void finishHandle(const Mantid::API::IAlgorithm *alg) override {
-    UNUSED_ARG(alg);
-    emit algCompletedSignal();
-  }
-};
-
-class FindPeakStrategyGeneric {
-public:
-  virtual void initialise(const std::string &wsName, const int workspaceIndex, const std::string &peakListName,
-                          const int FWHM, AlgorithmFinishObserver *obs) = 0;
-  virtual void execute() = 0;
-  virtual size_t peakNumber() const = 0;
-  virtual double getPeakCentre(const size_t peakIndex) const = 0;
-  virtual double getPeakHeight(const size_t peakIndex) const = 0;
-  virtual double getPeakWidth(const size_t peakIndex) const = 0;
-  virtual ~FindPeakStrategyGeneric(){};
-};
-
-template <typename T> class FindPeakStrategy : public FindPeakStrategyGeneric {
-public:
-  size_t peakNumber() const override { return m_peakCentres->size(); };
-  double getPeakCentre(const size_t peakIndex) const override { return m_peakCentres->operator[](peakIndex); };
-  double getPeakHeight(const size_t peakIndex) const override { return m_peakHeights->operator[](peakIndex); };
-  double getPeakWidth(const size_t peakIndex) const override { return m_peakWidths->operator[](peakIndex); };
-
-protected:
-  std::string m_peakListName;
-  std::unique_ptr<T> m_peakCentres;
-  std::unique_ptr<T> m_peakHeights;
-  std::unique_ptr<T> m_peakWidths;
-};
-
-class FindPeakConvolveStrategy : public QObject, public FindPeakStrategy<std::vector<double>> {
-  Q_OBJECT
-public:
-  void initialise(const std::string &wsName, const int workspaceIndex, const std::string &peakListName, const int FWHM,
-                  AlgorithmFinishObserver *obs) override;
-  void execute() override;
-
-private:
-  int m_FWHM;
-  AlgorithmFinishObserver *m_obs;        // Non-owning
-  MantidQt::API::AlgorithmDialog *m_dlg; // Non-owning
-};
-
-class FindPeakDefaultStrategy : public FindPeakStrategy<Mantid::API::ColumnVector<double>> {
-public:
-  void initialise(const std::string &wsName, const int workspaceIndex, const std::string &peakListName, const int FWHM,
-                  AlgorithmFinishObserver *obs) override;
-  void execute() override;
-
-private:
-  Mantid::API::IAlgorithm_sptr m_alg;
-};
 
 class PropertyHandler;
 class SelectFunctionDialog;
